refactor: unify stableStringify implementations
